Fix LogURIGenerator on Windows
Checking / is not the right way to distinguish an absolute path in
Windows.

Fixes #5786.
