+++ cio/io.go
@@ -18,7 +18,6 @@ package cio
 
 import (
 	"context"
-	"errors"
 	"fmt"
 	"io"
 	"net/url"
@@ -302,14 +301,21 @@ func LogFile(path string) Creator {
 // LogURIGenerator is the helper to generate log uri with specific scheme.
 func LogURIGenerator(scheme string, path string, args map[string]string) (*url.URL, error) {
 	path = filepath.Clean(path)
-	if !strings.HasPrefix(path, "/") {
-		return nil, errors.New("absolute path needed")
+	if !filepath.IsAbs(path) {
+		return nil, fmt.Errorf("%q must be absolute", path)
 	}
 
-	uri := &url.URL{
-		Scheme: scheme,
-		Path:   path,
+	// Without adding / here, C:\foo\bar.txt will become file://C:/foo/bar.txt
+	// which is invalid. The path must have three slashes.
+	//
+	// https://learn.microsoft.com/en-us/archive/blogs/ie/file-uris-in-windows
+	// > In the case of a local Windows file path, there is no hostname,
+	// > and thus another slash and the path immediately follow.
+	p := filepath.ToSlash(path)
+	if !strings.HasPrefix(path, "/") {
+		p = "/" + p
 	}
+	uri := &url.URL{Scheme: scheme, Path: p}
 
 	if len(args) == 0 {
 		return uri, nil
